moved setup utils to independent file
